Plugging in code for a WeaponBase.
This class is intended to be inherited by all other weapons in the game and will contain the base functionality common between weapon types., including things like hitscan, projectile, and melee types with a number of configurable default properties.

I have also gone and moved around (again) some of the C++ files which should be a more solid and readable project structure.
Everything should be somewhat like this

|-- Cerberus
    |-- <Folder Name>
        |-- Public
            |-- <File>.h
        |-- Private
            |-- <File>.cpp