More work on WeaonBase
Working on hitscan capabilities, still needs to be tested out. But I created BP_Weapon in the Dev folder for displaying the skeletal mesh and attempting to set some default functionality thats been opened up by the WeaponBase C++ class.

Also spawns the weapon into the players hand via the character blueprint. May wanna do that in C++ in the future but for now it does what it needs to