Clearly the Dev environment setup tools werent sufficient - This should be better
Basically removed the buildsource.bat and just placed everything inside of Setup.bat. When running that it will prompt to select your engine build location. Once selected it will generate  .uepath file which will then be used (if the UE5.sln does not exist) to run the Setup.bat and the GenerateProjectFiles.bat in the engine directory, then compile all the engine code. On any following runs of this new Setup.bat file, it will just search for the .ue5path file and compile engine code (So everything should only be done once)