Finished the general infrastructure for attributes, with working health
The CerberusWorldGameMode handles setting the player state, and the controller uses that to help initialize the pawn. Once the playerstate is initialized it creates an AbilitySystemComponent that gets used to drive the HealthComponent and its health attribute. Other attributes and components can be used in the same way.
I have also tied the health in blueprints to a working HUD. The blueprints for that is inside of the the character blueprint.