Configuration and upgrading the outlining for interactable objects. Can now be any color, and object specific.
a PostEditChangeProperty function was also added for the editor to be able to preview the change in color. This helps because the new OutlineColorValue property is what controls this, and values between 1-255 can be mapped to whatever color you wish. For example currently 255 = White, 254 = Red etc.
Setting the value 0 will disable the outline.