Moving around of files, networking fixes for new mesh, and preview actor creation
CerberusPreviewActor is a base class that is supposed to be used with blueprints to create UI previews of actors such as characters or weapons.
Fixed a networking issue related to the use of the modular character model, where the camera on the server and client werent seeing the same things. This was fixed by not doing separate calculations on both, but validating the client calculations were correct
Adding of equipable items. Items can now be created in blueprints and equipped to your character.
Other UI updates