Made the PreviewActor class a lot my dynamic, requiring much less setup
No longer need to create a blueprint manually setting up all the SkeletalMeshComp's and tying it together with the EEquipableSlot enum. Now its done automatically.
Blueprints will still need to be made for 2 reasons -- First reason, is that the actor that needs a preview needs to have a reference of a valid class to use. This is default editable in the character class currently. Second reason is for setting up animations on the preview actor. But all other legwork is done.